app/pyproject.toml
[tool.poetry]
name = "roomcounter"
version = "0.1.0"
description = "Web-Application for counting people in rooms"
authors = ["Dario Götz <dario.goetz@googlemail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = {version = "^1.6.1", extras = ["dotenv"]}
fastapi = "^0.61.1"
uvicorn = "^0.11.8"
python-multipart = "^0.0.5"
jinja2 = "^2.11.2"
sqlalchemy = "^1.3.19"
python-jose = "^3.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
